crud-api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
pymysql==1.1.0
pydantic[email]>=2.7.4
cryptography>=42.0.0
python-dotenv==1.0.1
